requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.0.6
django-crispy-forms==1.14.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2==2.9.5
python-decouple==3.6
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.2.0
